example/fastapi-enum/pyproject.toml
[project]
name = "fastapi-enum"
version = "0.1.0"
description = ""
authors = [
    {name = "Wilson Santos",email = "wilson.santos.dev87@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pydantic-enum @ ../../../pydantic-enum",
]

[tool.poetry]
packages = [{include = "fastapi_enum", from = "src"}]

# [tool.poetry.dependencies]
# pydantic-enum = {path = "../../../pydantic-enum/", develop = true}

[tool.poetry.dependencies]
pydantic-enum = {develop = true}

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
